requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
websockets==12.0
motor==3.3.2
redis==5.0.1
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1
python-dotenv==1.0.1
pydantic==2.6.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9
gunicorn==21.2.0 